RAG_app/backend/requirements.txt
chromadb==1.3.4
fastapi==0.121.1
protobuf==5.29.5
pydantic==2.12.4
uvicorn==0.38.0
google-generativeai==0.8.5
